requirements.txt
Flask==1.1.1
gunicorn
tensorflow==2.3.0
numpy===1.16.0
opencv-python==4.4.0.46
Pillow==7.0.0